Replace relative links to base URL in rich texts
The relative links used in Prismic `https://{{relative}}` is now
replaced to the current store's base URL within rich texts. This
was already the case for relative links in the "link" element, but
this has been copied to rich texts as well.

Also the double slash after the base URL is solved. It is common
to write a link in Prismic as `https://{{relative}}/something` which
is changed to `https://www.mydomain.com//something`, adding an
unwanted extra slash, because the base URL already contains a trailing
slash.