Implement dynamic breadcrumbs
To display the title of the Prismic pages in the breadcrumbs, they
are now loaded automatically from the code. To make that happen,
the `Block\Wrapper` needs to be used and will automatically add the
breadcrumbs.

This needs to be made nicer, so that will e a todo for later on.